refactor: manager widget 파일 경로 변경
